refactor(artifact-provenance): extract version writer
